Fix dangling pointer bug in ExtendedOptionData
The ExtendedOptionData structure was previously allocated on the stack
in american_option_price(), but its address was stored in the PDE solver's
callbacks. This created a dangling pointer that could be accessed after
the stack frame was destroyed.

Changes:
- Allocate ExtendedOptionData dynamically with malloc
- Add internal_data field to AmericanOptionResult to track allocated memory
- Implement american_option_free_result() to properly clean up resources
- Update all call sites to use american_option_free_result() instead of pde_solver_destroy()

This fixes the test CallRightBoundaryConditionNonZero which was failing
due to the boundary condition function receiving garbage x_max values
from the corrupted ExtendedOptionData.